Add unary local operations to local_operation lib
