+++ source/framework/algorithm/include/lue/framework/algorithm/cos.hpp
@@ -2,64 +2,33 @@
 #include "lue/framework/algorithm/policy/all_values_within_domain.hpp"
 #include "lue/framework/algorithm/policy/default_policies.hpp"
 #include "lue/framework/algorithm/policy/default_value_policies.hpp"
-#include "lue/framework/algorithm/unary_local_operation.hpp"
-#include <cmath>
+#include "lue/framework/partitioned_array.hpp"
 
 
 namespace lue {
-namespace detail {
+    namespace policy::cos {
 
-template<
-    typename InputElement>
-class Cos
-{
+        template<
+            typename Element>
+        using DefaultPolicies = policy::DefaultPolicies<
+            AllValuesWithinDomain<Element>,
+            OutputElements<Element>,
+            InputElements<Element>>;
 
-public:
+        template<
+            typename Element>
+        using DefaultValuePolicies = policy::DefaultValuePolicies<
+            AllValuesWithinDomain<Element>,
+            OutputElements<Element>,
+            InputElements<Element>>;
 
-    static_assert(std::is_floating_point_v<InputElement>);
+    }  // namespace policy::cos
 
-    using OutputElement = InputElement;
-
-    OutputElement operator()(
-        InputElement const& input_element) const noexcept
-    {
-        return std::cos(input_element);
-    }
-
-};
-
-}  // namespace detail
-
-
-namespace policy::cos {
 
     template<
-        typename Element>
-    using DefaultPolicies = policy::DefaultPolicies<
-        AllValuesWithinDomain<Element>,
-        OutputElements<Element>,
-        InputElements<Element>>;
-
-    template<
-        typename Element>
-    using DefaultValuePolicies = policy::DefaultValuePolicies<
-        AllValuesWithinDomain<Element>,
-        OutputElements<Element>,
-        InputElements<Element>>;
-
-}  // namespace policy::cos
-
-
-template<
-    typename Element,
-    Rank rank>
-PartitionedArray<Element, rank> cos(
-    PartitionedArray<Element, rank> const& array)
-{
-    using Functor = detail::Cos<Element>;
-    using Policies = policy::cos::DefaultPolicies<Element>;
-
-    return unary_local_operation(Policies{}, array, Functor{});
-}
+        typename Element,
+        Rank rank>
+    PartitionedArray<Element, rank> cos(
+        PartitionedArray<Element, rank> const& array);
 
 }  // namespace lue
