+++ source/framework/algorithm/include/lue/framework/algorithm/definition/logical.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include "lue/framework/algorithm/definition/logical_and.hpp"
 #include "lue/framework/algorithm/definition/logical_inclusive_or.hpp"
-// TODO #include "lue/framework/algorithm/definition/logical_not.hpp"
+#include "lue/framework/algorithm/definition/logical_not.hpp"
