+++ source/framework/algorithm/include/lue/framework/algorithm/downstream_distance.hpp
@@ -1,89 +1,9 @@
 #pragma once
-#include "lue/framework/algorithm/flow_direction.hpp"
-#include "lue/framework/algorithm/unary_local_operation.hpp"
 #include "lue/framework/algorithm/policy.hpp"
-#include <hpx/serialization.hpp>
-#include <cmath>
+#include "lue/framework/partitioned_array.hpp"
 
 
 namespace lue {
-    namespace detail {
-
-        template<
-            typename FlowDirectionElement,
-            typename DistanceElement>
-        class DownstreamDistance
-        {
-
-            public:
-
-                static_assert(std::is_floating_point_v<DistanceElement>);
-
-                using OutputElement = DistanceElement;
-
-
-                DownstreamDistance()=default;
-
-
-                DownstreamDistance(
-                    DistanceElement const& cell_size):
-
-                    _cell_size{cell_size},
-                    _cell_size_diagonal{std::sqrt(DistanceElement{2}) * cell_size}
-
-                {
-                }
-
-
-                DistanceElement operator()(
-                    FlowDirectionElement const flow_direction) const noexcept
-                {
-                    DistanceElement distance{0};
-
-                    if(
-                        flow_direction == north_west<FlowDirectionElement> ||
-                        flow_direction == north_east<FlowDirectionElement> ||
-                        flow_direction == south_west<FlowDirectionElement> ||
-                        flow_direction == south_east<FlowDirectionElement>)
-                    {
-                        distance = _cell_size_diagonal;
-                    }
-                    else if(
-                        flow_direction == north<FlowDirectionElement> ||
-                        flow_direction == west<FlowDirectionElement> ||
-                        flow_direction == east<FlowDirectionElement> ||
-                        flow_direction == south<FlowDirectionElement>)
-                    {
-                        distance = _cell_size;
-                    }
-
-                    return distance;
-                }
-
-
-            private:
-
-                friend class hpx::serialization::access;
-
-
-                template<typename Archive>
-                void serialize(
-                    Archive& archive,
-                    [[maybe_unused]] unsigned int const version)
-                {
-                    archive & _cell_size & _cell_size_diagonal;
-                }
-
-
-                DistanceElement _cell_size;
-
-                DistanceElement _cell_size_diagonal;
-
-        };
-
-    }  // namespace detail
-
-
     namespace policy::downstream_distance {
 
         template<
@@ -113,14 +33,7 @@ namespace lue {
     PartitionedArray<DistanceElement, rank> downstream_distance(
         Policies const& policies,
         PartitionedArray<FlowDirectionElement, rank> const& flow_direction,
-        DistanceElement const cell_size)
-    {
-        static_assert(rank == 2);
-
-        using Functor = detail::DownstreamDistance<FlowDirectionElement, DistanceElement>;
-
-        return unary_local_operation(policies, flow_direction, Functor{cell_size});
-    }
+        DistanceElement const cell_size);
 
 
     template<
