+++ source/framework/algorithm/include/lue/framework/algorithm/log.hpp
@@ -1,36 +1,11 @@
 #pragma once
-#include "lue/framework/algorithm/unary_local_operation.hpp"
 #include "lue/framework/algorithm/policy/all_values_within_domain.hpp"
 #include "lue/framework/algorithm/policy/default_policies.hpp"
 #include "lue/framework/algorithm/policy/default_value_policies.hpp"
-#include <cmath>
+#include "lue/framework/partitioned_array.hpp"
 
 
 namespace lue {
-    namespace detail {
-
-        template<
-            typename InputElement>
-        class Log
-        {
-
-            public:
-
-                static_assert(std::is_floating_point_v<InputElement>);
-
-                using OutputElement = InputElement;
-
-                OutputElement operator()(
-                    InputElement const& input_element) const noexcept
-                {
-                    return std::log(input_element);
-                }
-
-        };
-
-    }  // namespace detail
-
-
     namespace policy::log {
 
         template<
@@ -73,10 +48,7 @@ namespace lue {
         Rank rank>
     PartitionedArray<Element, rank> log(
         Policies const& policies,
-        PartitionedArray<Element, rank> const& array)
-    {
-        return unary_local_operation(policies, array, detail::Log<Element>{});
-    }
+        PartitionedArray<Element, rank> const& array);
 
 
     template<
