+++ source/framework/algorithm/src/local_operation/cast.cpp
@@ -1,0 +1,1 @@
+#include "lue/framework/algorithm/definition/cast.hpp"
