+++ source/framework/algorithm/src/local_operation/cos.cpp
@@ -1,0 +1,1 @@
+#include "lue/framework/algorithm/definition/cos.hpp"
