+++ source/framework/algorithm/src/local_operation/logical_not.cpp
@@ -1,0 +1,1 @@
+#include "lue/framework/algorithm/definition/logical_not.hpp"
