+++ source/framework/algorithm/src/local_operation/sin.cpp
@@ -1,0 +1,1 @@
+#include "lue/framework/algorithm/definition/sin.hpp"
