+++ source/framework/algorithm/test/cast_test.cpp
@@ -1,7 +1,7 @@
 #define BOOST_TEST_MODULE lue framework algorithm cast
 #include "lue/framework/algorithm/all.hpp"
 #include "lue/framework/algorithm/array_like.hpp"
-#include "lue/framework/algorithm/cast.hpp"
+#include "lue/framework/algorithm/definition/cast.hpp"
 #include "lue/framework/algorithm/definition/equal_to.hpp"
 #include "lue/framework/algorithm/create_partitioned_array.hpp"
 #include "lue/framework/test/array.hpp"
