+++ source/framework/algorithm/test/log_test.cpp
@@ -2,7 +2,7 @@
 #include "lue/framework/algorithm/all.hpp"
 #include "lue/framework/algorithm/create_partitioned_array.hpp"
 #include "lue/framework/algorithm/definition/equal_to.hpp"
-#include "lue/framework/algorithm/log.hpp"
+#include "lue/framework/algorithm/definition/log.hpp"
 #include "lue/framework/test/array.hpp"
 #include "lue/framework/test/hpx_unit_test.hpp"
